Add movieDbController for queries to DB and movieApiController for request to movie API, and update both controllers to display DB rating on MovieInfo page.
